+++ src/GraphQL/Language/AST/Directive.cs
@@ -8,17 +8,30 @@ namespace GraphQL.Language.AST
     /// </summary>
     public class Directive : AbstractNode
     {
+        /// <summary>
+        /// Initializes a new instance of a directive node with the specified parameters.
+        /// </summary>
         public Directive(NameNode node)
         {
             NameNode = node;
         }
 
+        /// <summary>
+        /// Returns the name of this directive.
+        /// </summary>
         public string Name => NameNode.Name;
 
+        /// <summary>
+        /// Returns the <see cref="NameNode"/> which contains the name of this directive.
+        /// </summary>
         public NameNode NameNode { get; set; }
 
+        /// <summary>
+        /// Returns the node containing a list of argument nodes for this directive.
+        /// </summary>
         public Arguments Arguments { get; set; }
 
+        /// <inheritdoc/>
         public override IEnumerable<INode> Children
         {
             get { yield return Arguments; }
@@ -27,6 +40,9 @@ namespace GraphQL.Language.AST
         /// <inheritdoc />
         public override string ToString() => $"Directive{{name='{Name}',arguments={Arguments}}}";
 
+        /// <summary>
+        /// Compares this instance with another instance by name.
+        /// </summary>
         protected bool Equals(Directive other)
         {
             if (other == null)
@@ -35,6 +51,7 @@ namespace GraphQL.Language.AST
             return string.Equals(Name, other.Name, StringComparison.InvariantCulture);
         }
 
+        /// <inheritdoc/>
         public override bool IsEqualTo(INode obj)
         {
             if (obj is null)
