+++ src/GraphQL/Language/AST/Document.cs
@@ -3,10 +3,16 @@ using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Represents the root node of a document.
+    /// </summary>
     public class Document : AbstractNode
     {
         private readonly List<IDefinition> _definitions;
 
+        /// <summary>
+        /// Initializes a new instance with no children.
+        /// </summary>
         public Document()
         {
             _definitions = new List<IDefinition>();
@@ -14,14 +20,27 @@ namespace GraphQL.Language.AST
             Fragments = new Fragments();
         }
 
+        /// <summary>
+        /// Gets or sets the query before being parsed into an AST document.
+        /// </summary>
         public string OriginalQuery { get; set; }
 
+        /// <inheritdoc/>
         public override IEnumerable<INode> Children => _definitions;
 
+        /// <summary>
+        /// Returns a list of operation nodes for this document.
+        /// </summary>
         public Operations Operations { get; }
 
+        /// <summary>
+        /// Returns a list of fragment nodes for this document.
+        /// </summary>
         public Fragments Fragments { get; }
 
+        /// <summary>
+        /// Adds a <see cref="FragmentDefinition"/> or <see cref="Operation"/> node to this document.
+        /// </summary>
         public void AddDefinition(IDefinition definition)
         {
             _definitions.Add(definition ?? throw new ArgumentNullException(nameof(definition)));
@@ -43,6 +62,7 @@ namespace GraphQL.Language.AST
         /// <inheritdoc />
         public override string ToString() => $"Document{{definitions={string.Join(", ", _definitions)}}}";
 
+        /// <inheritdoc/>
         public override bool IsEqualTo(INode node)
         {
             if (node is null)
