+++ src/GraphQL/Language/AST/Fields.cs
@@ -3,6 +3,9 @@ using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Represents a list of field nodes within a document.
+    /// </summary>
     public class Fields : IEnumerable<Field>
     {
         private readonly Dictionary<string, Field> _fields;
@@ -12,8 +15,14 @@ namespace GraphQL.Language.AST
             _fields = fields;
         }
 
+        /// <summary>
+        /// Returns a new instance that contains no field nodes.
+        /// </summary>
         public static Fields Empty() => new Fields(new Dictionary<string, Field>());
 
+        /// <summary>
+        /// Adds a field node to the list.
+        /// </summary>
         public void Add(Field field)
         {
             var name = field.Alias ?? field.Name;
@@ -28,6 +37,7 @@ namespace GraphQL.Language.AST
             }
         }
 
+        /// <inheritdoc/>
         public IEnumerator<Field> GetEnumerator() => _fields.Values.GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
