+++ src/GraphQL/Language/AST/IDefinition.cs
@@ -1,5 +1,8 @@
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Represents a fragment definition node or an operation node.
+    /// </summary>
     public interface IDefinition : INode
     {
     }
