+++ src/GraphQL/Language/AST/IFragment.cs
@@ -1,5 +1,8 @@
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Represents a fragment spread node or an inline fragment node.
+    /// </summary>
     public interface IFragment : ISelection
     {
     }
