+++ src/GraphQL/Language/AST/INode.cs
@@ -19,6 +19,7 @@ namespace GraphQL.Language.AST
 
         /// <summary>
         /// Determines if the node is equal to another node.
+        /// This typically returns <see langword="true"/> if the node type and the node name matches.
         /// </summary>
         bool IsEqualTo(INode node);
     }
