+++ src/GraphQL/Language/AST/ISelection.cs
