+++ src/GraphQL/Language/AST/IType.cs
@@ -1,5 +1,8 @@
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Represents a type node within a document. This could be a named type, list type, or non-null type.
+    /// </summary>
     public interface IType : INode
     {
     }
