+++ src/GraphQL/Language/AST/NullValue.cs
@@ -1,12 +1,16 @@
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Represents the value null within a document.
+    /// </summary>
     public class NullValue : AbstractNode, IValue
     {
         object IValue.Value => null;
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public override string ToString() => "null";
 
+        /// <inheritdoc/>
         public override bool IsEqualTo(INode obj)
         {
             if (obj is null)
