+++ src/GraphQL/Language/AST/SourceLocation.cs
@@ -1,7 +1,13 @@
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Provides information regarding the location of a node within a document's original query text.
+    /// </summary>
     public class SourceLocation
     {
+        /// <summary>
+        /// Initializes a new instance with the specified parameters.
+        /// </summary>
         public SourceLocation(int line, int column, int start = -1, int end = -1)
         {
             Line = line;
@@ -10,22 +16,38 @@ namespace GraphQL.Language.AST
             End = end;
         }
 
+        /// <summary>
+        /// Returns the start position within the query text.
+        /// </summary>
         public int Start { get; }
 
+        /// <summary>
+        /// Returns the end position within the query text.
+        /// </summary>
         public int End { get; }
 
+        /// <summary>
+        /// Returns the line number within the query text.
+        /// </summary>
         public int Line { get; }
 
+        /// <summary>
+        /// Returns the column number within the query text.
+        /// </summary>
         public int Column { get; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public override string ToString() => $"line={Line}, column={Column}, start={Start}, end={End}";
 
+        /// <summary>
+        /// Compares this instance to another based on the line and column values.
+        /// </summary>
         protected bool Equals(SourceLocation other)
         {
             return Line == other.Line && Column == other.Column;
         }
 
+        /// <inheritdoc/>
         public override bool Equals(object obj)
         {
             if (obj is null)
@@ -37,12 +59,7 @@ namespace GraphQL.Language.AST
             return Equals((SourceLocation)obj);
         }
 
-        public override int GetHashCode()
-        {
-            unchecked
-            {
-                return (Line * 397) ^ Column;
-            }
-        }
+        /// <inheritdoc/>
+        public override int GetHashCode() => unchecked((Line * 397) ^ Column);
     }
 }
