+++ src/GraphQL/Language/AST/Variables.cs
@@ -5,10 +5,16 @@ using System.Linq;
 
 namespace GraphQL.Language.AST
 {
+    /// <summary>
+    /// Contains a list of variables (name &amp; value tuples) that have been gathered from the document and attached <see cref="Inputs"/>.
+    /// </summary>
     public class Variables : IEnumerable<Variable>
     {
         private List<Variable> _variables;
 
+        /// <summary>
+        /// Adds a variable to the list.
+        /// </summary>
         public void Add(Variable variable)
         {
             if (variable == null)
@@ -20,13 +26,16 @@ namespace GraphQL.Language.AST
             _variables.Add(variable);
         }
 
+        /// <summary>
+        /// Returns the first variable with a matching name, or <see langword="null"/> if none are found.
+        /// </summary>
         public object ValueFor(string name)
         {
             var variable = _variables?.FirstOrDefault(v => v.Name == name);
             return variable?.Value;
         }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IEnumerator<Variable> GetEnumerator()
         {
             if (_variables == null)
@@ -37,37 +46,8 @@ namespace GraphQL.Language.AST
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public override string ToString() => _variables?.Count > 0 ? $"Variables{{{string.Join(", ", _variables)}}}" : "Variables(Empty)";
     }
 
-    public class VariableDefinitions : IEnumerable<VariableDefinition>
-    {
-        private List<VariableDefinition> _variables;
-
-        public void Add(VariableDefinition variable)
-        {
-            if (variable == null)
-                throw new ArgumentNullException(nameof(variable));
-
-            if (_variables == null)
-                _variables = new List<VariableDefinition>();
-
-            _variables.Add(variable);
-        }
-
-        /// <inheritdoc />
-        public IEnumerator<VariableDefinition> GetEnumerator()
-        {
-            if (_variables == null)
-                return Enumerable.Empty<VariableDefinition>().GetEnumerator();
-
-            return _variables.GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-
-        /// <inheritdoc />
-        public override string ToString() => _variables?.Count > 0 ? $"VariableDefinitions{{{string.Join(", ", _variables)}}}" : "VariableDefinitions(Empty)";
-    }
 }
